Close a potential loophole that could have resulted in missed events.
A poll 901 seconds before an event would result in a 900-second sleep. If the
next calendar query results in a few seconds of internet latency (not unusual
in the least), this could result in the event being completely missed. So, in
the last half hour of the countdown, just sleep five minutes at a time.