Put a more useful title during alerts
While playing an alert sound, instead of showing the name of the alert file
being played, show the event name. The alert file is great for debugging the
firing of alerts, but the event name is much more important information for
regular usage, once the alerts are firing perfectly every time.